Remove PaginationListView
